Refactor: test 디렉터리 리팩터링
