+++ 이지은/freelec-springboot2-webservice/src/test/java/com/jojoldu/book/springboot/web/HelloControllerTest.java
@@ -1,12 +1,10 @@
-package com.jojoldu.book.springboot;
+package com.jojoldu.book.springboot.web;
 
-import com.jojoldu.book.springboot.web.HelloController;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
 import org.springframework.test.context.junit.jupiter.SpringExtension;
-import org.springframework.test.context.junit4.SpringRunner;
 import org.springframework.test.web.servlet.MockMvc;
 
 import static org.hamcrest.Matchers.is;
