+++ 이지은/freelec-springboot2-webservice/src/test/java/com/jojoldu/book/springboot/web/dto/HelloResponseDtoTest.java
@@ -1,4 +1,4 @@
-package com.jojoldu.book.springboot.dto;
+package com.jojoldu.book.springboot.web.dto;
 
 import com.jojoldu.book.springboot.web.dto.HelloResponseDto;
 import org.junit.jupiter.api.Test;
